rpi/requirements.txt
minimalmodbus>=2.0.1
RPi.GPIO>=0.7.1
